Organize specs into directories
