+++ spec/crm/crm_data_spec.rb
@@ -1,4 +1,4 @@
-require_relative "../lib/crm/crm_data"
+require_relative "../../lib/crm/crm_data"
 
 describe CRM::Data do
